chore: remove dead code reported by deadcode
golang.org/x/tools/cmd/deadcode flagged 8 unreachable functions
across the tree. Each one had test callers but no production
caller; removing them and their now-orphaned tests cuts ~260
lines without changing any live behavior.

Removed (with their tests):
- internal/gitx: ClassifyGitError + ErrorClass type + Err*
  constants — replaced by gitx.ClassifyError(err) string, which
  vcs.gitErrorClassifier uses.
- internal/config: LastUpdated() — unused timestamp helper.
- internal/tui: renderRow() — production code calls
  renderStyledRow(); plain renderRow only existed for tests.
- cmd/repokeeper/edit: validateUpstreamRef — orphaned from a
  prior edit flow.
- cmd/repokeeper/import: removeRegistryEntryByRepoID — only
  reached from a coverage-boost test.

Moved (test-only helpers, kept under _test.go to drop them from
the shipped binary):
- cloneImportedRepos / cloneImportedReposWithProgress /
  cloneImportedEntriesWithProgress now live in
  cmd/repokeeper/import_helpers_test.go. They collapse the
  production plan/execute pair into one call for tests.

Verified: go build ./..., go test ./..., golangci-lint v2.11.4,
deadcode ./... — all clean.
