+++ internal/gitx/parse.go
@@ -93,58 +93,3 @@ func ParseRevListCount(output string) (int, int) {
 	behind, _ := strconv.Atoi(strings.TrimSpace(parts[1]))
 	return ahead, behind
 }
-
-// ClassifyError categorizes a git command error into a known error type.
-type ErrorClass string
-
-const (
-	ErrAuth     ErrorClass = "auth"
-	ErrNetwork  ErrorClass = "network"
-	ErrNoRemote ErrorClass = "no_remote"
-	ErrCorrupt  ErrorClass = "corrupt"
-	ErrNotARepo ErrorClass = "not_a_repo"
-	ErrTimeout  ErrorClass = "timeout"
-	ErrUnknown  ErrorClass = "unknown"
-)
-
-// ClassifyGitError inspects the error output and returns a classification.
-func ClassifyGitError(stderr string) ErrorClass {
-	lower := strings.ToLower(stderr)
-
-	switch {
-	case strings.Contains(lower, "authentication failed"),
-		strings.Contains(lower, "permission denied"),
-		strings.Contains(lower, "could not read from remote"),
-		strings.Contains(lower, "invalid credentials"):
-		return ErrAuth
-
-	case strings.Contains(lower, "could not resolve host"),
-		strings.Contains(lower, "connection refused"),
-		strings.Contains(lower, "network is unreachable"),
-		strings.Contains(lower, "connection timed out"),
-		strings.Contains(lower, "unable to access"),
-		strings.Contains(lower, "unable to connect"):
-		return ErrNetwork
-
-	case strings.Contains(lower, "no remote repository"),
-		strings.Contains(lower, "no such remote"):
-		return ErrNoRemote
-
-	case strings.Contains(lower, "object file is empty"),
-		strings.Contains(lower, "loose object"),
-		strings.Contains(lower, "corrupt"),
-		strings.Contains(lower, "bad object"):
-		return ErrCorrupt
-
-	case strings.Contains(lower, "not a git repository"):
-		return ErrNotARepo
-
-	case strings.Contains(lower, "deadline exceeded"),
-		strings.Contains(lower, "timed out"),
-		strings.Contains(lower, "timeout"):
-		return ErrTimeout
-
-	default:
-		return ErrUnknown
-	}
-}
