+++ internal/gitx/parse_test.go
@@ -170,45 +170,3 @@ var _ = Describe("ParseRevListCount", func() {
 		Expect(behind).To(Equal(10))
 	})
 })
-
-var _ = Describe("ClassifyGitError", func() {
-	It("detects auth errors", func() {
-		Expect(gitx.ClassifyGitError("fatal: Authentication failed")).To(Equal(gitx.ErrAuth))
-	})
-
-	It("detects permission denied", func() {
-		Expect(gitx.ClassifyGitError("Permission denied (publickey)")).To(Equal(gitx.ErrAuth))
-	})
-
-	It("detects network errors", func() {
-		Expect(gitx.ClassifyGitError("fatal: unable to access: Could not resolve host")).To(Equal(gitx.ErrNetwork))
-	})
-
-	It("detects connection refused", func() {
-		Expect(gitx.ClassifyGitError("fatal: unable to connect: Connection refused")).To(Equal(gitx.ErrNetwork))
-	})
-
-	It("detects no remote", func() {
-		Expect(gitx.ClassifyGitError("fatal: No remote repository specified")).To(Equal(gitx.ErrNoRemote))
-	})
-
-	It("detects no such remote", func() {
-		Expect(gitx.ClassifyGitError("fatal: No such remote 'origin'")).To(Equal(gitx.ErrNoRemote))
-	})
-
-	It("detects corrupt repo", func() {
-		Expect(gitx.ClassifyGitError("error: object file is empty")).To(Equal(gitx.ErrCorrupt))
-	})
-
-	It("detects not a repo", func() {
-		Expect(gitx.ClassifyGitError("fatal: not a git repository")).To(Equal(gitx.ErrNotARepo))
-	})
-
-	It("detects timeout", func() {
-		Expect(gitx.ClassifyGitError("context deadline exceeded")).To(Equal(gitx.ErrTimeout))
-	})
-
-	It("returns unknown for unrecognized", func() {
-		Expect(gitx.ClassifyGitError("some random error")).To(Equal(gitx.ErrUnknown))
-	})
-})
